Update patch set 3
Patch Set 3:

It would be nice to get this right. If we were to factor our this association logic to a final pass, when we know everything, we could do it right because we could tell which predeclared types were locally overridden, couldn't we?

It seems like this would be a simple loop containing this code, factored out, and called after the package has been read.

Patch-set: 3
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>