Update patch set 10
Patch Set 10:

> > `go get A` will use exactly the dependencies that `A` was published with.
> 
> It will use exactly the dependencies that A was published with, but not the dependencies that were used to run its CI tests, or that you get if you run "go build" inside that module. I have a feeling that that discrepancy will turn out to be a significant gotcha.

If CI tests use only one view of the module, it should probably be the outside view. See https://golang.org/issue/24666.

If a module doesn't build from the outside, then it also can't be built from within any other module — at least, not without applying patches in that other module, and those patches can break other module dependencies.

And for the few cases where unpublished changes really are necessary, users always have the option to execute the build from within a module — or even from within *that* module in the module cache.


> > (Remember, modules are designed to produce reproducible builds!) If you want to pick up patch releases you have to run `go get -u=patch A`, in which case you should also be able to combine a specific (fixed or earlier-than-broken) version of the dependency:
> > 
> >     go get -u=patch A C@version-that-fixes-regression
> > 
> > or
> > 
> >     go get -u=patch A B@version-that-works-with-c
> 
> Yes, I agree that it's possible to do this, but this requires that the person doing the install knows exactly what module to replace with what exact version. Wouldn't it be nice if the maintainer of the module containing the command had a way to encapsulate that information in the source tree somehow? For example by including a replace statement in the go.mod file? :)

If a large fraction of binaries can't build (or don't work) using committed, mainline versions of their dependencies, that would mean that the ecosystem is in a very bad state (and we should do something more drastic to fix it, not just hack around it).

Patch-set: 10