Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > > `go get A` will use exactly the dependencies that `A` was published with.
> > 
> > It will use exactly the dependencies that A was published with, but not the dependencies that were used to run its CI tests, or that you get if you run "go build" inside that module. I have a feeling that that discrepancy will turn out to be a significant gotcha.
> 
> If CI tests use only one view of the module, it should probably be the outside view. See https://golang.org/issue/24666.
> 
> If a module doesn't build from the outside, then it also can't be built from within any other module — at least, not without applying patches in that other module, and those patches can break other module dependencies.

I agree that's a problem. But I'm not convinced that it's a problem for main packages, which are probably the principal use case for this CL.

Main packages act as roots of the module tree. Building a main package within another module isn't really a design goal AFAIK. Yes, there's a hacky way to add a main package as a dependency to another package, but I'm not convinced it's good practice to do so.

If there's any time that it's appropriate to perform replacements and exclusions, it's surely when building a non-test binary!

An issue with go get is that it conflates two use cases: building a top level binary and pulling in a package as a dependency. When there isn't a current module, I'm not sure that the latter use case is that important any more - we have `go mod download` for that.

The issue is made harder by the fact that a command like `go get $module/...` does both jobs at once, and there's no way to tell what the intention of the user is.

Certainly for myself, when I use `go get` outside of a module, it's almost always because I want to install some binary or other. One possibility might be to recognize that case and make `go get` ignore non-main packages (and error if only non-main packages are specified) when there's no current module. Another might be to add a flag that causes `go get` to act in "binary install" mode. That's something I've pondered in the past, as it happens, because I consider it somewhat problematic that `go get $module/...` might or might not have arbitrary side-effects on my $GOBIN.

If we know when we're installing binaries, then the choice becomes simpler, I think: apply exclusions and replacements when installing binaries (because transitive dependencies aren't so important then) and not otherwise.

> And for the few cases where unpublished changes really are necessary, users always have the option to execute the build from within a module — or even from within *that* module in the module cache.
> 
> 
> > > (Remember, modules are designed to produce reproducible builds!) If you want to pick up patch releases you have to run `go get -u=patch A`, in which case you should also be able to combine a specific (fixed or earlier-than-broken) version of the dependency:
> > > 
> > >     go get -u=patch A C@version-that-fixes-regression
> > > 
> > > or
> > > 
> > >     go get -u=patch A B@version-that-works-with-c
> > 
> > Yes, I agree that it's possible to do this, but this requires that the person doing the install knows exactly what module to replace with what exact version. Wouldn't it be nice if the maintainer of the module containing the command had a way to encapsulate that information in the source tree somehow? For example by including a replace statement in the go.mod file? :)
> 
> If a large fraction of binaries can't build (or don't work) using committed, mainline versions of their dependencies, that would mean that the ecosystem is in a very bad state (and we should do something more drastic to fix it, not just hack around it).

I wasn't necessarily talking about a "large fraction" of binaries. I'm worried about a situation where even a very small fraction of modules need to publish custom install instructions for a Go binary in a security alert situation because any exclude and replace statements within the binary's module are ignored.

I agree about the usefulness of https://github.com/golang/go/issues/24666. That would provide a way to turn off replace/exclude when there is a current module; but I think that the flip side of that is also important: to be able to turn *on* replace/exclude when there *isn't* a current module. If 24666 is fixed (for example by adding a -no-module flag) and `go get` applies replace/exclude by default, then both use cases are satisfied, I think.

To quote Russ:

> Minimal version selection is very simple. It achieves simplicity by
> eliminating all flexibility about what the answer must be: the build
> list is exactly the versions specified in the requirements. A real system
> needs more flexibility, for example the ability to exclude certain module
> versions or replace others.

I worry that by placing the burden of specifying exclusions and replacements entirely on the client, what we're doing is removing much of that essential flexibility, because clients will in practice never use it.

Patch-set: 10