Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> I think this package is basically frozen, in favor of the cryptobyte package.

I'm not sure I understand what the motivation here is.  There is an "encodings" module that contains various de/serialization mechanisms.  It seems entirely consistent to have a general use case asn1 package in there for de/serializing data to/from ASN.1.

Compared to the ASN.1 implementation that's bundled with crypto, I would expect that the encoding/asn1 package be much more applicable to general use cases while the crypto implementation would specifically support only the design goals of the crypto package, it's corner cases and nothing else.  

In fact, I would even go so far as to expect that proposals extending the functionality of the ASN1 package within crypto, in a way that was not required by crypto, would be rejected.  Why complicate a package beyond its purpose.

If the strategic intention is to consolidate onto a single ASN.1 library, then wouldn't it make more sense to deprecate the crypto implementation, and refactor crypto to leverage encoding/asn1.  

In addition, having looked at the cryptobyte doc, it looks far less friendly and/or scalable to use in general use cases with large and dynamic data structures compared to encodings/asn1.  The included examples are based on very simple data structures, and yet much harder to read than examples of encoding/asn1 that use bigger structures (IMO).

Patch-set: 1
CC: Gerrit User 30182 <30182@62eb7196-b449-3ce5-99f1-c037f21e1705>