cmd/vendor,cmd/internal/objfile: update to golang.org/x/arch@5a4828b
This updates master to fix the ppc64 objdump. There were many cases where
the Go objdump was generating opcodes that didn't exist in the Go assembler,
or generated operands in the wrong order. The goal is to generate a Go
objdump that is acceptable to the Go assembler, or as close as possible.

In addition to the changes to golang.org/x/arch, changes were made to disasm.go
to provide consistent handling of an instruction containing all 0s. This also adds
a comment displaying the corresponding gnu output to the right of the Go objdump
output for ppc64. For example:

  0x11024       e8610050          MOVD 80(R1),R3             // ld r3,80(r1)
  0x11028       f8610020          MOVD R3,32(R1)             // std r3,32(r1)

Change-Id: Ie8d2d534e13b9a64852c99b4b864a9c08ed7e036