Update patch set 7
Patch Set 7: Run-TryBot+1

LGTM.

Now the only real changes (compared to the prior code) are that we check for isPredeclared (and exclude those); and to make that work correctly, functions are processed after everything else.

Which is exactly the minimal change we needed to make this work right.

Thanks for bearing with me.

Patch-set: 7
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1