cmd/go/internal/lockedfile: fix EDEADLK error on aix/ppc64
On AIX, fcntl calls with L_SETLKW can return EDEADLK if several process
try to read several files in common and will result on a deadlock if
the current process waits until a lock is released by another process.
This commit puts to sleep the current process if such error is returned
before trying to call fcntl again.

Change-Id: Ifd862c39d12e544f9f4621697d5be07b1a6492e1