Update patch set 1
Patch Set 1:

A regression test would be helpful, and ideally a link to an upstream kernel bug.

As I understand it, EDEADLK should only fire if the locks acquired by the processes involved actually do deadlock — that is, if all of the threads of all of the processes involved are waiting on locks — and if that were to happen we really would want the operation to fail (instead of just sitting around waiting on the lock).

In contrast, after cmd/go acquires a given lock it should be able to make progress on some thread, even if another thread is blocked. I don't believe that we ever acquire multiple simultaneous file locks from the same goroutine, although goroutine-multiplexing does imply that we might acquire a lock from one thread and release it from another.

If EDEADLK is firing for things that are not actually deadlocks — for example, if one thread in each process is blocked on acquiring the other's lock but other threads are not blocked — that's arguably a kernel bug, but if it's a common kernel bug perhaps we should work around it anyway.

Patch-set: 1
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>