cmd/go/internal/work: always chown directory in TestRespectSetgidDir
On BSD, os.Chmod fails if the SetGID bit is requested but the user is
not a member of the directory's group. On Linux, os.Chmod silently
fails to set the bit in that case, which renders the rest of the test
invalid. Avoid both by calling os.Chown unconditionally.

os.Chown doesn't work on Windows or Plan 9, but since the SetGID bit
is ignored on those platforms the test is irrelevant anyway. Skip it.

Fixes #29160

Change-Id: Ifefe9136ec6de4739bb3bbde396fb1e5a8d79882