Update patch set 3
Patch Set 3:

There are still a number of references to godoc in https://tip.golang.org/doc/effective_go.html. Is this intentional? 

The first reference to godoc is the second paragraph under "Commentary":

> The program—and web server—godoc processes Go source files to extract documentation about the contents of the package. 

Then: 

> The nature and style of these comments determines the quality of the documentation godoc produces.

... and others.

It seems to me that we should consistently refer to go doc throughout, not least because of the comparisons drawn with gofmt and the examples being command-line based. 

We could/should add a short paragraph explaining how godoc relates to go doc, is a webserver etc. 

At the moment it doesn't feel particularly consistent.

Patch-set: 3
CC: Gerrit User 16375 <16375@62eb7196-b449-3ce5-99f1-c037f21e1705>