cmd/compile: fix length overflow when appending elements to a slice
Instead of testing len(slice)+numNewElements > cap(old) use
uint(len(slice)+numNewElements) > uint(cap(old)) to test
if a slice needs to be grown in an append operation.

This prevents a possible overflow when len(slice) is near the maximum
int value and the addition of the constant number of new elements
make it overflow and wrap around to a negative number which compares
as smaller than the capacity of the slice.

Appending a slice to a slice with append(s1, s2...) already used
a uint comparison to test slice capacity and therefore was not
vurnerable to the same overflow issue.

Fixes: #29190

Change-Id: I41733895838b4f80a44f827bf900ce931d8be5ca