runtime: don't coalesce scavenged spans with unscavenged spans
As a result of changes earlier in Go 1.12, the scavenger became much
more aggressive. In particular, when scavenged and unscavenged spans
coalesced, they would always become scavenged. This resulted in most
spans becoming scavenged over time. While this is good for keeping the
RSS of the program low, it also causes many more undue page faults and
many more calls to madvise. For example, newly-mapped memory would
likely coalesce with an scavenged span, which results in unnecessary
page faults.

For most applications, the impact of this was negligible. But for
applications that repeatedly grow and shrink the heap by large amounts,
the overhead can be significant. The overhead was especially obvious on
older versions of Linux where MADV_FREE isn't available and
MADV_DONTNEED must be used.

This change makes it so that scavenged spans will never coalesce with
unscavenged spans. This solves the issues with newly-mapped memory, and
results in fewer page faults overall. Aside from this, the expected
impact of this change is more heap growths on average, as span
allocations will be less likely to be fulfilled. To mitigate this
slightly, this change also coalesces spans eagerly after scavenging, to
at least ensure that all scavenged spans and all unscavenged spans are
coalesced with each other.

Update #14045.

Change-Id: I21ac0018a8b048ee1c575e1af0ef55d561131fe2