runtime: don't coalesce scavenged spans with unscavenged spans
Prior to this change, when scavenged and unscavenged spans would
coalesce, they would always become scavenged. This results in most spans
becoming scavenged over time. While this is good for keeping the RSS of
the program low, it also causes many more undue page faults. While page
faults in general aren't great, with MADV_FREE they weren't quite so bad
because the OS kicks out pages lazily. With MADV_DONTNEED this was much
more of a problem. Notably, newly-mapped memory would likely coalesce
with an scavenged span, which results in definitely unnecessary page
faults.

This change makes it so that scavenged spans will never coalesce with
unscavenged spans. This solves the issues with newly-mapped memory, and
fewer page faults overall. Aside from this, the expected impact of this
change is more heap growths on average, as span allocations will be less
likely to be fulfilled. To mitigate this slightly, this change also
coalesces spans eagerly after scavenging, to at least ensure that all
scavenged spans and all unscavenged spans are coalesced with each other.

Change-Id: I21ac0018a8b048ee1c575e1af0ef55d561131fe2