unicode: update Unicode tables using new generator
This is now using the range table generator from x/text.

This change is in preparation for Unicode 11.0.0. The old
code generated incorrect tables for this new version.
Tables are now more compact as a side-effect of this change.

Because the package code does not allow any imports from
outside the standard library, the generation code now
lives in x/text.

Change-Id: I0d0afbf5389b5de3c965331cde8ee2978f51b7ea