os: don't let sendFile put a pipe into blocking mode
If we pass a pipe to sendFile, it will call the Fd method, which on
Unix systems will put the pipe into blocking mode, which means that
deadlines will stop working. Don't let that happen. We can't use sendfile
with a pipe anyhow, so the only effect is to avoid going into blocking mode.

This coincidentally already worked on everything but GNU/Linux.

Fixes #28330

Change-Id: If093417a225fe44092bd2c0dbbc3937422e98c0b