Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I'm not sure about hard-coding this to linux/arm64. This is both too restrictive (people could be running the tests on a non-containerized linux/arm64) and too permissive (people could be running containerized on non-linux/arm64). Maybe the test should just skip if unshare fails with EPERM?

I considered that, but I'm worried that over time all the builders end up running in some kind of environment and unshare isn't available. Then, the test could regress but it would take a while for everyone to notice. The advantage of this solution though is that if another builder's configuration changes in such a way that the test can't run, then it'll be immediately obvious.

Though, I'm not very happy with this solution either, since the underlying issue isn't a product of linux/arm64 specifically, it's a product of how our builders are configured.

I guess I can have the testprog return something like "skip" and just check for that string specifically if it fails with EPERM? It doesn't really address my concern about the test not getting run properly over time, but at least if that ever does become an issue it's much more explicit and obvious as to why the test is passing.

For lack of a better solution I'll do that, since I don't feel too strongly either way. I'm also open to suggestions. :)

Patch-set: 2
Reviewer: Gerrit User 25391 <25391@62eb7196-b449-3ce5-99f1-c037f21e1705>