Update patch set 1
Patch Set 1:

I won't vote against this because I don't want a fight, but I wish it didn't go in as it encourages ugly I/O patterns that until now were only available through the syscall package, where they belong (if anywhere). Promoting this to the os package endorses them, and that's what I object to.

So although I won't formally vote, I just wanted to go on the record. I will add this comment to the issue.

I miss Unix. It was a nice operating system.

Patch-set: 1
CC: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>