Update patch set 2
Patch Set 2:

The title was reverted to the original issue related to the display of default vet flags.
The previous patch created a merge conflict.
A simpler solution displays the flags when using f.i. go help testflag.
They appear as:
"...checks believed to be always worth addressing, i.e.
-atomic -bool -buildtags -nilfunc -printf
..."

The "\t    " is to respect the format of the rest of the help text.

Patch-set: 2
Reviewer: Gerrit User 26957 <26957@62eb7196-b449-3ce5-99f1-c037f21e1705>