go/printer: generalize bench harness, add unbalanced benchmark
This change generalizes the existing go/printer benchmarking harness
to allow adding new benchmarks, and it adds a new benchmark for the
formatting of huge, unbalanced binary expressions (for Issue #23350).

To avoid adding what would be a ~60kb testfile to the repository, the
.go file with the big binary expression to be formatted for the new
benchmark is generated at runtime, and then deleted.

Change-Id: I4d851685e7701aae6191c851c72ec0f5bcc5bc49