Update patch set 6
Patch Set 6:

(2 comments)

> Do the Inline marks potentially prevent some stores from moving around?

No. They take a memory but don't produce one, so they don't restrict scheduling at all.
In fact, it's not really necessary that they are in any particular place in the instruction stream. They could all be at the end of the function and it would still work.
I've placed them inline so they appear at a natural location in the code, and maybe as such they'd be a good breakpoint for a debugger.

Patch-set: 6
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>