Update patch set 2
Patch Set 2: Code-Review+2

OR (really OGR) is a 4-byte instruction as is this MOVD replacement (which compiles to a LGR instruction). I think there was a reason when we did the initial port that it needed to be 4 bytes (PC-line table maybe?) but that reason might no longer exist so we could try replacing MOVD with the NOP instruction (which is 2 bytes long). That would avoid the change to the codegen tests.

Happy to keep this CL as-is for 1.12 though.

Patch-set: 2
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2