Update patch set 5
Patch Set 5:

> > > OR (really OGR) is a 4-byte instruction as is this MOVD replacement (which compiles to a LGR instruction). I think there was a reason when we did the initial port that it needed to be 4 bytes (PC-line table maybe?) but that reason might no longer exist so we could try replacing MOVD with the NOP instruction (which is 2 bytes long). That would avoid the change to the codegen tests.
> > 
> > Just checked. The nop still needs to be 4 bytes.
> > 
> 
> What was the underlying reason?

I made a mistake. I assumed obj.ANOP would encode to a 2-byte NOP but of course it doesn't. So my experiment was invalid since the NOP I was adding was omitted from the instruction stream.

Looking at the uses of ginsnop it is probably fine to use the normal 2-byte NOP, though there is no easy way to do that without adding a new instruction. I'll put together a CL for 1.13. I think what we have now is fine for 1.12.

Patch-set: 5